requirements.txt
opencv-python==4.5.5.64
flask==2.0.3
pyserial==3.5
numpy==1.22.0
